feat: #122 영업기회 파이프라인 관련 api 추가, lead 금액 관련 컬럼 Long 타입으로 수정, contract랑 estimate 매핑 설정

+++ src/main/java/beyond/samdasoo/lead/dto/LeadRequestDto.java
@@ -10,9 +10,9 @@ import java.time.LocalDate;
 public class LeadRequestDto {
     private String name;      // 리드 이름
     private LeadStatus status;  // 리드 상태
-    private int expSales;     // 예상 매출
+    private Long expSales;     // 예상 매출
     private int expMargin;    // 예상 이익률
-    private int expProfit;    // 예상 이익금
+    private Long expProfit;    // 예상 이익금
     private Long process;      // 프로세스
     private Long subProcess;   // 서브 프로세스
     private int successPer;   // 성공 확률
